server/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
boto3==1.34.109
python-dotenv==1.0.1
cryptography==41.0.7
Flask-JWT-Extended==4.5.3
bcrypt==4.1.2
paramiko==3.4.0
